# contracts/openapi/inspector-api.v1.yaml
openapi: 3.0.3
info:
  title: Инспектор ИИ — REST API
  version: 0.24.0
  description: |
    Публичный REST API сервиса «Инспектор ИИ» (веб-клиент и внешние ИС, целевая — ИАИС «РиН»).
    Асинхронная pull-модель: загрузка → process_id → опрос статуса → получение протокола.

    Требования и их ID — docs/TZ.md.

    Соглашения:
    - все поля snake_case; идентификаторы — UUID (кроме params.id — integer, как в ТЗ);
    - время — RFC 3339 (UTC);
    - bbox — [x0, y0, x1, y1] в долях [0;1] видимой страницы (после CropBox/Rotate),
      начало координат — левый верхний угол (как viewport pdf.js);
    - страницы нумеруются с 1;
    - ошибки — объект Error.

    Изменения 0.2.0: пары страниц для экрана сравнения (PagePair, GET /protocols/{id}/page-pairs),
    compliance_percent в сводке, auto_start при загрузке, метод гипотез VISUAL_DIFF, фильтры findings по file_id/evidence_page,
    Finding.page_pair_id и rationale_source.

    Изменения 0.3.0: манифест комплекта при загрузке
    (UploadManifest), external_id объекта, серверный импорт папки (POST /documents/import),
    правка доказательств новой версией (POST /findings/{id}/evidence, evidence_history),
    массовое решение CONFIRM/CLARIFY по одному параметру (POST /findings/bulk-decision),
    PATCH /files — только выбор авторитетной редакции (метаданные не редактируются),
    сводка комплектности (ProcessInfo.completeness).

    Изменения 0.4.0 (Приложение 1 и «Перечень ИД» ред. 1.1): коды параметров M-001…M-132;
...

...
  license:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/files/{file_id}/signature:
    parameters:
      - $ref: '#/components/parameters/FileId'
    post:
      tags: [documents]
      operationId: attachSignature
      summary: Приложить файл электронной подписи к документу
      description: >-
        Подвязка под ЭЦП: открепленная подпись (.sig, .p7s, .sgn) принимается, сохраняется рядом
        с документом и видна в карточке файла. **Криптографическая проверка не выполняется** —
        `verification` всегда `NOT_VERIFIED`, а `signature_status` документа от этого не меняется:
        он приходит из реестра комплекта. Проверка подписи и статусы `VALID` / `INVALID`
        появятся, когда будет доступ к УКЭП и службе проверки (REQ-INT-02 — сертификатов на
        соревновании не дают). Повторная загрузка заменяет приложенный файл.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [signature]
              properties:
                signature:
                  type: string
                  format: binary
      responses:
        '200':
          description: Карточка файла с приложенной подписью
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          description: Файл подписи больше допустимого
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/documents/upload:
    post:
      tags: [documents]
      operationId: uploadDocuments
      summary: Загрузка/дозагрузка документов (REQ-UPL-01..08)
      description: |
        Без process_id — создаётся новая проверка (статус PENDING).
        С process_id — дозагрузка в существующую проверку (запрещено в PARSING и FINALIZED → 409).
        Лимиты: файл ≤ 50 МБ, пакет ≤ 200 МБ; форматы PDF, DOCX, XML; антивирус до сохранения.
        Отклонённые файлы возвращаются в files[] со status=REJECTED и error; если превышен лимит
        пакета — отклоняется весь пакет (413).

        Комплект целиком (0.15.0):
        - архив .zip в files[] раскрывается: содержимое — отдельными файлами, стадия — по папкам внутри
          («Проектная документация» и т.п.), сам архив в комплект не входит. Предел архива — предел пакета.
          Не распаковался — REJECTED с CORRUPTED_FILE или FILE_TOO_LARGE и причиной в message;
        - папка: имя файла в multipart может быть путём («Комплект/Рабочая документация/АР.pdf»,
          webkitRelativePath браузера). В original_name — последний сегмент, стадия — по папкам пути;
          .zip внутри папки — карточка плюс содержимое, DWG и прочие форматы из папки или архива —
          карточкой (OTHER / SKIPPED), как при серверном импорте;
        - реестр (registry.* / реестр* / manifest* / document_manifest.*, CSV/XLSX/JSON/JSONL) ближе всего
          к корню архива или папки используется как реестр комплекта и документом не считается.
          Поля registry и manifest формы важнее найденного.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadForm'
      responses:
        '202':
          description: Файлы приняты (частично или полностью)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          description: Превышен общий лимит пакета (200 МБ) — REQ-UPL-05
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/processes/{process_id}/registry:
    parameters:
      - $ref: '#/components/parameters/ProcessId'
    post:
      tags: [documents]
      operationId: uploadRegistry
      summary: Загрузка или замена реестра файлов комплекта
      description: |
        Реестр (CSV, XLSX или JSON) заменяет ранее загруженный: метаданные файлов берутся из него,
        ожидаемый состав пересчитывается. Если по проверке уже есть протокол (или auto_start=true) и сейчас нет обработки,
        запускается повторное сравнение — появится новая версия протокола. Запрещено в PARSING и FINALIZED (409).
        Роли: INSPECTOR, SUPERVISOR, ADMIN.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                registry:
                  type: string
                  format: binary
                manifest:
                  type: string
                  description: Реестр JSON-строкой по схеме UploadManifest (если нет файла registry)
                auto_start:
                  type: string
                  enum: ['true', 'false']
                  default: 'true'
      responses:
        '200':
          description: Реестр применён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistryApplyResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/protocols/{protocol_id}/export:
    parameters:
      - $ref: '#/components/parameters/ProtocolId'
    get:
      tags: [protocols]
      operationId: exportProtocol
      summary: Экспорт протокола в PDF / DOCX / XML / JSON / GOLD / формат ответа организаторов (модуль 7)
      description: |
        Отражает текущее состояние решений инспектора. Для финализированного протокола файл сохраняется
        в хранилище (protocols/{process_id}/v{n}-{время финализации}.{ext}) и повторно не формируется.
      parameters:
        - name: format
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ExportFormat'
      responses:
        '200':
          description: Файл протокола
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Protocol'
                  - $ref: '#/components/schemas/GoldExport'
                  - $ref: '#/components/schemas/SubmissionExport'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mtls:
      type: http
      scheme: bearer
      description: >-
        Для внешних ИС — взаимная TLS-аутентификация клиентским сертификатом (УКЭП, REQ-INT-02).
        OpenAPI 3.0 не имеет типа mutualTLS, поэтому схема описана текстом.
  parameters:
...
    ProcessId:
      name: process_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FileId:
      name: file_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProtocolId:
      name: protocol_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Некорректный запрос (валидация OpenAPI)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Не аутентифицирован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Операция недопустима в текущем статусе
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    DocStage:
      type: string
      description: Стадия документации — ПД / РД / ИД
      enum: [PD, RD, ID]
...
    ApprovalStatus:
      type: string
      description: |
        Статус утверждения редакции (реестр файлов, «Перечень ИД» ред. 1.1):
        DRAFT — черновик, эталоном не служит; APPROVED — утверждена; FOR_CONSTRUCTION — выдана «В производство работ»;
        SUPERSEDED — заменена новой редакцией (хранится для аудита, из эталонного сравнения исключается);
        CANCELLED — аннулирована (исключается); UNKNOWN — реестра нет и ML не определил статус.
      enum: [DRAFT, APPROVED, FOR_CONSTRUCTION, SUPERSEDED, CANCELLED, UNKNOWN]
...
    ExportFormat:
      type: string
      description: |
        json — Protocol как есть; pdf / docx — печатная форма (REQ-CMP-08); xml — та же структура для внешних ИС;
        gold — выгрузка по схеме GOLD (Приложение 1), по записи на атомарный finding и гипотезу;
        submission — ответ участника по submission_schema.json организаторов (коды параметров вида KR-055).
      enum: [json, pdf, docx, xml, gold, submission]
...
    FileFormat:
      type: string
      description: OTHER — формат без анализа (архивы, DWG, изображения); принимается карточкой файла при серверном импорте и из загруженных папки или архива
      enum: [PDF, DOCX, XML, OTHER]
...
    MetadataSource:
      type: string
      description: Откуда взяты метаданные файла (приоритет — манифест, затем ML, затем имя файла)
      enum: [MANIFEST, ML, FILENAME]
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Машиночитаемый код (например UNSUPPORTED_FORMAT, INVALID_STATUS)
        message:
          type: string
          description: Сообщение для пользователя (рус.)
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
...
    UploadForm:
      type: object
      required: [files]
      properties:
        object_id:
          type: string
          format: uuid
          description: Обязателен, если не передан process_id
        process_id:
          type: string
          format: uuid
          description: Для дозагрузки в существующую проверку
        auto_start:
          type: string
          enum: ['true', 'false']
          default: 'true'
          description: >-
            'true' — сразу запустить анализ после загрузки (для новой проверки — полный,
            для дозагрузки — инкрементальный). 'false' — только загрузить (запуск через /start).
        registry:
          type: string
          format: binary
          description: >-
            Машиночитаемый реестр файлов комплекта (CSV, XLSX или JSON) — обязателен по «Перечню ИД» ред. 1.1.
            Без реестра комплект принимается со статусом полноты CLARIFICATION_REQUIRED.
            Колонки — docs/domain/registry.md.
        manifest:
          type: string
          description: >-
            Реестр в виде JSON-строки по схеме UploadManifest (альтернатива полю registry).
            Поля реестра важнее оценки по имени файла и ML.
        stage_hints:
          type: string
          description: >-
            JSON-объект {"<имя файла>": "PD"|"RD"|"ID"} — подсказка стадии от пользователя
            (ML может уточнить по содержимому)
        files:
          type: array
          description: >-
            Документы (PDF, DOCX, XML), архивы .zip с комплектом или файлы папки с путём в имени
            («Проектная документация/ПЗ.pdf») — см. описание операции.
          items:
            type: string
            format: binary
...
    RegistryApplyResult:
      type: object
      required: [process_id, entries_total, matched_files, completeness, analysis_started]
      properties:
        process_id:
          type: string
          format: uuid
        registry_file_name:
          type: string
          nullable: true
        entries_total:
          type: integer
        matched_files:
          type: integer
          description: Сколько загруженных файлов получили метаданные из реестра
        completeness:
          $ref: '#/components/schemas/CompletenessSummary'
        analysis_started:
          type: boolean
          description: Анализ перезапущен, чтобы новая версия протокола учла реестр
    UploadResponse:
      type: object
      required: [process_id, files, upload_status]
      properties:
        process_id:
          type: string
          format: uuid
        process_status:
          $ref: '#/components/schemas/ProcessStatus'
        files:
          type: array
          items:
            $ref: '#/components/schemas/UploadedFileResult'
        upload_status:
          type: array
          description: Ровно по одному статусу на стадию (PD_*, RD_*, ID_*)
          items:
            $ref: '#/components/schemas/StageUploadStatus'
...
    FileInfo:
      type: object
      required: [id, object_id, process_id, original_name, format, size_bytes, sha256, processing_status, uploaded_at]
      properties:
        id:
          type: string
          format: uuid
        object_id:
          type: string
          format: uuid
        process_id:
          type: string
          format: uuid
        original_name:
          type: string
        format:
          $ref: '#/components/schemas/FileFormat'
        size_bytes:
          type: integer
        sha256:
          type: string
        doc_stage:
          type: string
          allOf:
            - $ref: '#/components/schemas/DocStage'
          nullable: true
        doc_kind:
          type: string
          nullable: true
          description: Вид документа (например «АОСР», «Основной комплект КЖ», «Пояснительная записка»)
        discipline:
          type: string
          nullable: true
          description: Раздел/марка (ПЗ, АР, КР, КЖ, ЭОМ, ВК, ОВ, СС, ИОС1...)
        document_code:
          type: string
          nullable: true
          description: Шифр документа
        revision:
          type: string
          nullable: true
          description: Номер изменения/редакции
        approval_status:
          $ref: '#/components/schemas/ApprovalStatus'
        approval_date:
          type: string
          format: date
          nullable: true
        predecessor_id:
          type: string
          format: uuid
          nullable: true
        successor_id:
          type: string
          format: uuid
          nullable: true
        is_authoritative:
          type: boolean
          nullable: true
          description: Ручной выбор авторитетной редакции инспектором (null — автоматически)
        metadata_confidence:
          type: number
          nullable: true
        metadata_source:
          $ref: '#/components/schemas/MetadataSource'
        authoritative_basis:
          type: string
          nullable: true
          description: Основание выбора авторитетной редакции инспектором
        external_file_id:
          type: string
          nullable: true
          description: file_id из реестра (например ALT79B-000015)
        in_registry:
          type: boolean
          description: Файл описан в реестре комплекта
        sheet_page_range:
          type: string
          nullable: true
        signature_status:
          $ref: '#/components/schemas/SignatureStatus'
        signature:
          description: Приложенный файл открепленной подписи; null — не прикладывали
          type: object
          allOf:
            - $ref: '#/components/schemas/SignatureAttachment'
          nullable: true
        duplicate_of:
          type: string
          format: uuid
          nullable: true
        excluded_from_comparison:
          type: boolean
          description: Файл хранится для аудита, но в эталонное сравнение не идёт
        exclusion_reason:
          type: string
          nullable: true
          description: Дубликат, заменённая/аннулированная редакция, исключён инспектором
        processing_status:
          $ref: '#/components/schemas/FileProcessingStatus'
        quality:
          $ref: '#/components/schemas/FileQuality'
        error:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
          description: Заголовок документа (из реестра; позже — по штампу)
        # --- сведения о документе от ML (0.20.0): null — не определено или файл ещё не разобран.
        # Размер файла — size_bytes выше, его не дублируем
        project_code:
          type: string
          nullable: true
          description: Шифр проекта — часть шифра документа до стадии и марки
          example: ЖС-РД-270121
        language:
          type: string
          enum: [ru, en, mixed]
          nullable: true
          description: Язык текста по доле кириллицы
        scan_share:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
          description: Доля страниц, прошедших распознавание (скан)
        pdf_version:
          type: string
          nullable: true
          example: PDF 1.7
        pdf_producer:
          type: string
          nullable: true
          description: Программа, записавшая PDF
          example: AutoCAD PDF
        pdf_creator:
          type: string
          nullable: true
          description: Программа, в которой создан документ
          example: Revit
        encrypted:
          type: boolean
          nullable: true
          description: PDF зашифрован
        developer_org:
          type: string
          nullable: true
          description: >-
            Организация-разработчик ПД/РД по штампу или титулу (0.21.0). Для ИД null; у ИП только «ИП» —
            без фамилии (152-ФЗ)
          example: ООО «ГЕНПРОЕКТ»
        storage_key:
          type: string
          description: >-
            Ключ файла в хранилище (озеро данных): raw/{sha256}. Одинаковое содержимое хранится один раз,
            сам файл в БД не лежит — только ссылка и метаданные
          example: raw/9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08
        uploaded_by:
          type: string
          format: uuid
          nullable: true
        uploaded_by_name:
          type: string
          nullable: true
          description: ФИО загрузившего пользователя
        uploaded_at:
          type: string
          format: date-time
...
    SubmissionExport:
      type: object
      description: >-
        Ответ участника в формате организаторов (submission_schema.json из датасета, разбор —
        data/samples/dataset-overview.md). Оцениваются файл и страница доказательства, bbox не нужен.
      required: [object_id, checks]
      properties:
        object_id:
          type: string
          description: Идентификатор объекта у заказчика (ObjectInfo.external_id), иначе внутренний id
        checks:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionCheck'

    # ----------------------------------------------------------------- admin
...